% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{bestBlastF1Score}
\alias{bestBlastF1Score}
\title{Function to compute the performance F1-scores of Best Blast protein function
predictions using the Gene Ontology. The Best Blast method simply passes the
GO Term annotations found for the best Blast Hit to the respective query
protein.}
\usage{
bestBlastF1Score(best.blast.res.tbl, blast.hit.annos,
  bbrt.query.col = getOption("MapMan2GO.best.blast.tbl.query.col",
  "query.ukb.short.id"),
  bbrt.hit.col = getOption("MapMan2GO.best.blast.tbl.hit.col",
  "hit.ukb.short.id"),
  bha.prot.col = getOption("MapMan2GO.blast.hit.goas.prot.col", 3),
  bha.go.col = getOption("MapMan2GO.blast.hit.goas.go.col", 2),
  reference.gene.annos = getOption("MapMan2GO.reference.gene.annos",
  ukb.ref.goas), rga.gene.col = getOption("MapMan2GO.rga.gene.col", "V4"),
  rga.anno.col = getOption("MapMan2GO.rga.anno.col", "V2"),
  process.ref.annos.funk = getOption("MapMan2GO.process.ref.annos.funk",
  MapMan2GO::addAncestors))
}
\arguments{
\item{best.blast.res.tbl}{An instance of \code{data.frame} result of
invoking \code{MapMan2GO::extractBestBlastHits}.}

\item{blast.hit.annos}{An instance of \code{data.frame} holding the GO
annotations for the proteins in the Blast database. Must have at least two
columns, one with the Blast Hit protein identifier (expected to be in short
Uniprot format), and another holding the GO Term identifiers. See this
packages Vignette, section 'Prepare UniprotKB Gene Ontology Annotations
(GOA) for processing in R.'.}

\item{bbrt.query.col}{An integer or string indicating the column of argument
\code{best.blast.res.tbl} in which to find the query protein identifiers.
Default is \code{getOption('MapMan2GO.best.blast.tbl.query.col',
'query.ukb.short.id')}.}

\item{bbrt.hit.col}{An integer or string indicating the column of
argument \code{best.blast.res.tbl} in which to find the hit protein
identifiers. Default is \code{getOption('MapMan2GO.best.blast.tbl.hit.col',
'hit.ukb.short.id')}.}

\item{bha.prot.col}{An integer or string indicating the column of argument
\code{blast.hit.annos} in which to find the hit protein identifiers.
Default is \code{getOption('MapMan2GO.blast.hit.goas.prot.col', 3)}.}

\item{bha.go.col}{An integer or string indicating the column of argument
\code{blast.hit.annos} in which to find the annotated GO Term
identifiers. Default is \code{getOption('MapMan2GO.blast.hit.goas.go.col',
2)}.}

\item{reference.gene.annos}{An instance of \code{base::data.frame} holding the
reference annotations for the referene genes, i.e. the TRUTH the mercator
predictions are evaluated against. Default is
\code{getOption('MapMan2GO.reference.gene.annos', ukb.ref.goas)}.}

\item{rga.gene.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene identifiers. Default
is \code{getOption('MapMan2GO.rga.gene.col', 'V4')}.}

\item{rga.anno.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene annotations. Default
is \code{getOption('MapMan2GO.rga.anno.col', 'V2')}.}

\item{process.ref.annos.funk}{Function applied to each reference gene's
annotation vector. Set to \code{base::identity} if you whish to leave them
unchanged. Because we are dealing with Gene Ontology Term annotions the
default here is \code{getOption( 'MapMan2GO.process.ref.annos.funk',
MapMan2GO::addAncestors)}.}
}
\value{
An instance of \code{base::data.frame} with the following numeric
columns: precision, recall, true.pos, false.pos, f1.score, mcc, gene.
}
\description{
Function to compute the performance F1-scores of Best Blast protein function
predictions using the Gene Ontology. The Best Blast method simply passes the
GO Term annotations found for the best Blast Hit to the respective query
protein.
}
