% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{bestBlastPredictions}
\alias{bestBlastPredictions}
\title{Function to generate the annotations (GO Term predictions) for Best Blast.}
\usage{
bestBlastPredictions(best.blast.tbl, blast.hit.goa)
}
\arguments{
\item{best.blast.tbl}{The result of invoking function
\code{MapMan2GO::extractBestBlastHits}.}

\item{blast.hit.goa}{The result of reading in a pre-processed form of the
UniProt GOA tables. See Vignette for more details. Is expected to have at
least two columns, 'V3' holding the Blast Hit identifier matching
'query.ukb.short.id' in \code{best.blast.tbl} and 'V2' holding the Gene
Ontology Terms assigned to these respective Hits.}
}
\value{
An instance of \code{base::data.frame} with two columns: 'query'
holds the respective query proteins and 'GO' the GO predictions obtained by
the Best Blast method.
}
\description{
Function to generate the annotations (GO Term predictions) for Best Blast.
}
