% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{extractBestBlastHits}
\alias{extractBestBlastHits}
\title{Function to parse a Blastp output table, retain only the best non self Hits,
where query and hit are identical, also discarding any hits that match
protein identifiers given one per line in argument
\code{exclude.accessions.lines.path}. This function expects protein
identifiers to be in the UniprotKB format, eg. 'sp|Q54IF9|MYBG_DICDI'.}
\usage{
extractBestBlastHits(path.2.blast.tbl, exclude.accessions.lines.path)
}
\arguments{
\item{path.2.blast.tbl}{Valid file path to the Blastp output table.}

\item{exclude.accessions.lines.path}{Valid file path to the file in which
protein identifiers are stored that are to be excluded from the hits.
Expected format is the short protein ID from uniprot, eg. 'Q54IF9'.}
}
\value{
An instance of \code{data.frame} with 14 columns, two additional to
the original blast tabular output: \code{'query.ukb.short.id'} and
\code{'hit.ukb.short.id'}, which hold the short Uniprot protein accessions
of the Query and Hit proteins respectively, eg. 'Q54IF9'. The other 12
columns are obtained from the Blast tabular output.
}
\description{
Function to parse a Blastp output table, retain only the best non self Hits,
where query and hit are identical, also discarding any hits that match
protein identifiers given one per line in argument
\code{exclude.accessions.lines.path}. This function expects protein
identifiers to be in the UniprotKB format, eg. 'sp|Q54IF9|MYBG_DICDI'.
}
