% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{interProScanF1Scores}
\alias{interProScanF1Scores}
\title{Computes the F1-scores of InterProScan protein function predictions.}
\usage{
interProScanF1Scores(ipr.annos,
  reference.genes = getOption("MapMan2GO.reference.genes", ref.gene.ids),
  reference.gene.annos = getOption("MapMan2GO.reference.gene.annos",
  ukb.ref.goas), rga.gene.col = getOption("MapMan2GO.rga.gene.col", "V4"),
  rga.anno.col = getOption("MapMan2GO.rga.anno.col", "V2"),
  process.ref.annos.funk = getOption("MapMan2GO.process.ref.annos.funk",
  MapMan2GO::addAncestors))
}
\arguments{
\item{ipr.annos}{An instance of \code{base::data.frame} with at least the
following required columns: 'V2' holds the function annotions, and 'V4'
holds the gene identifiers.}

\item{reference.genes}{A character vector of gene identifiers forming the
complete set of reference genes ('gold standard').}

\item{reference.gene.annos}{An instance of \code{base::data.frame} holding the
reference annotations for the referene genes, i.e. the TRUTH the mercator
predictions are evaluated against. Default is
\code{getOption('MapMan2GO.reference.gene.annos', ukb.ref.goas)}.}

\item{rga.gene.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene identifiers. Default
is \code{getOption('MapMan2GO.rga.gene.col', 'V4')}.}

\item{rga.anno.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene annotations. Default
is \code{getOption('MapMan2GO.rga.anno.col', 'V2')}.}

\item{process.ref.annos.funk}{Function applied to each reference gene's
annotation vector. Set to \code{base::identity} if you whish to leave them
unchanged. Because we are dealing with Gene Ontology Term annotions the
default here is \code{getOption( 'MapMan2GO.process.ref.annos.funk',
MapMan2GO::addAncestors )}.}
}
\value{
An instance of \code{base::data.frame} with the following numeric
columns: precision, recall, true.pos, false.pos, f1.score, gene.
}
\description{
Computes the F1-scores of InterProScan protein function predictions.
}
