\name{MapMan2GO-package}
\alias{MapMan2GO-package}
\alias{MapMan2GO}
\docType{package}
\title{
\packageTitle{MapMan2GO}
}
\description{
\packageDescription{MapMan2GO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MapMan2GO}
\packageIndices{MapMan2GO}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MapMan2GO}

Maintainer: \packageMaintainer{MapMan2GO}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
