% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{mercatorF1Score}
\alias{mercatorF1Score}
\title{For each row in \code{mercator.annos} measures the respective F1-Score of the
MapMan-Bin annotation based on the reference annotations.}
\usage{
mercatorF1Score(mercator.annos,
  reference.gene.annos = getOption("MapMan2GO.reference.gene.annos",
  ukb.ref.goas), rga.gene.col = getOption("MapMan2GO.rga.gene.col", "V5"),
  rga.anno.col = getOption("MapMan2GO.rga.anno.col", "V2"),
  process.ref.annos.funk = getOption("MapMan2GO.process.ref.annos.funk",
  MapMan2GO::addAncestors))
}
\arguments{
\item{mercator.annos}{an instance of \code{base::data.frame} produced by
invoking function \code{MapMan2GO::readMercatorResultTable}.}

\item{reference.gene.annos}{An instance of \code{base::data.frame} holding the
reference annotations for the referene genes, i.e. the TRUTH the mercator
predictions are evaluated against. Default is
\code{getOption('MapMan2GO.reference.gene.annos', ukb.ref.goas)}.}

\item{rga.gene.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene identifiers. Default
is \code{getOption('MapMan2GO.rga.gene.col', 'V5')}.}

\item{rga.anno.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene annotations. Default
is \code{getOption('MapMan2GO.rga.anno.col', 'V2')}.}

\item{process.ref.annos.funk}{Function applied to each reference gene's
annotation vector. Set to \code{base::identity} if you whish to leave them
unchanged. Because we are dealing with Gene Ontology Term annotions the
default here is \code{getOption( 'MapMan2GO.process.ref.annos.funk',
MapMan2GO::addAncestors )}.}
}
\value{
An instance of \code{base::data.frame} with the folque gold standard
proteinslowing numeric columns: precision, recall, true.pos, false.pos,
f1.score, mcc, gene (the reference)
}
\description{
For each row in \code{mercator.annos} measures the respective F1-Score of the
MapMan-Bin annotation based on the reference annotations.
}
