% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{mutualInformationBinGoaGenesGoas}
\alias{mutualInformationBinGoaGenesGoas}
\title{Compute the mutual information between a MapMan Bin's GOA and the GOAs found
for it's reference genes. In this, regard each single GO Term as a event in
the sample space. Because each GO Term can only have binary count values in
the Bin's GOA the Bin's counts are normalized by multiplying with them with
the number of reference genes.}
\usage{
mutualInformationBinGoaGenesGoas(genes.goa)
}
\arguments{
\item{genes.goa}{A list of character vectors. Names are reference gene
identifiers and values, the character vectors, represent each gene's GO Term
Annotations. See \code{compoundGoAnnotationEntropy} for more details.}
}
\value{
A numeric value, the computed mutual information in bits (log2).
}
\description{
Compute the mutual information between a MapMan Bin's GOA and the GOAs found
for it's reference genes. In this, regard each single GO Term as a event in
the sample space. Because each GO Term can only have binary count values in
the Bin's GOA the Bin's counts are normalized by multiplying with them with
the number of reference genes.
}

