% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{performanceScores}
\alias{performanceScores}
\title{Compute the Matthew's correlation coefficient (mcc) and F1-Score for the
prediction \code{predicted.annos}. F1-Score as the harmonic mean of
precision and recall. mcc using \code{mccr::mccr}.}
\usage{
performanceScores(true.annos, predicted.annos,
  universe.annos = getOption("MapMan2GO.performance.universe.annotations",
  ukb.ref.universe.gos),
  na.for.empty.references = getOption("MapMan2GO.f1.na.for.empty.references",
  TRUE))
}
\arguments{
\item{true.annos}{A vector of atomic annotations representing the TRUTH.}

\item{predicted.annos}{A vector of atomic annotations representing the
PREDICTIONS.}

\item{universe.annos}{A vector of atomic annotations representing all
possible annotations that could theoretically be assigned, i.e. the
'annotation universe'. Default is
\code{getOption('MapMan2GO.performance.universe.annotations',
ukb.ref.universe.gos)}.}

\item{na.for.empty.references}{boolean indicating whether to return all NA
values in case no references are given. In the case of protein function
prediction missing references can be interpreted as missing knowledge,
because there are no proteins with no function. Default is
\code{getOption('MapMan2GO.f1.na.for.empty.references', TRUE)}.}
}
\value{
An instance of \code{base::data.frame} with the following numeric
columns: n.truth, n.pred, true.pos, false.pos, precision, recall,
false.pos.rate, f1.score, mcc
}
\description{
Compute the Matthew's correlation coefficient (mcc) and F1-Score for the
prediction \code{predicted.annos}. F1-Score as the harmonic mean of
precision and recall. mcc using \code{mccr::mccr}.
}
