% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{plotDistAsHistAndBox}
\alias{plotDistAsHistAndBox}
\title{Generates a two row plot with the first one being a Histogram and the secons
row a horizontal Boxplot.}
\usage{
plotDistAsHistAndBox(x, main)
}
\arguments{
\item{x}{The values passed into \code{hist} and \code{boxplot}}

\item{The}{main title of the resulting plot}
}
\value{
TRUE if and only if no error has occurred
}
\description{
Generates a two row plot with the first one being a Histogram and the secons
row a horizontal Boxplot.
}
