% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{predictorPerformance}
\alias{predictorPerformance}
\title{Computes the performance scores of protein function predictions.}
\usage{
predictorPerformance(pred.annos, pa.gene.col, pa.anno.col,
  reference.genes = getOption("MapMan2GO.reference.genes", ref.gene.ids),
  reference.gene.annos = getOption("MapMan2GO.reference.gene.annos",
  ukb.ref.goas), rga.gene.col = getOption("MapMan2GO.rga.gene.col", "V4"),
  rga.anno.col = getOption("MapMan2GO.rga.anno.col", "V2"),
  process.predicted.annos.funk = getOption("MapMan2GO.process.predicted.annos.funk",
  identity), process.annos.funk = getOption("MapMan2GO.process.annos.funk",
  MapMan2GO::addAncestors))
}
\arguments{
\item{pred.annos}{An instance of \code{base::data.frame} with at least two
columns, one holding the protein identifier and the other the predicted
annotations.}

\item{pa.gene.col}{An integer or string indicating in which column of
\code{pred.annos} to lookup the gene identifiers.}

\item{pa.anno.col}{An integer or string indicating in which column of
\code{pred.annos} to lookup the predicted annotations.}

\item{reference.genes}{A character vector of gene identifiers forming the
complete set of reference genes ('gold standard').}

\item{reference.gene.annos}{An instance of \code{base::data.frame} holding the
reference annotations for the referene genes, i.e. the TRUTH the mercator
predictions are evaluated against. Default is
\code{getOption('MapMan2GO.reference.gene.annos', ukb.ref.goas)}.}

\item{rga.gene.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene identifiers. Default
is \code{getOption('MapMan2GO.rga.gene.col', 'V4')}.}

\item{rga.anno.col}{An integer or string indicating in which column of
\code{reference.gene.annos} to lookup the reference gene annotations. Default
is \code{getOption('MapMan2GO.rga.anno.col', 'V2')}.}

\item{process.predicted.annos.funk}{Function applied to the predicted
annotations before being passed to argument \code{process.annos.funk}. Set
to \code{base::identity} (the default), if you want to leave them unchanged.
Set to \code{MapMan2GO::splitMapManBinGOAs} if the function prediction
method was Mercator / MapMan2GO. Default is
\code{getOption('MapMan2GO.process.predicted.annos.funk', identity)}.}

\item{process.annos.funk}{Function applied to each
annotation vector. Set to \code{base::identity} if you whish to leave them
unchanged. Because we are dealing with Gene Ontology Term annotions the
default here is \code{getOption( 'MapMan2GO.process.annos.funk',
MapMan2GO::addAncestors )}.}
}
\value{
An instance of \code{base::data.frame} with the columns as returned
by \code{MapMan2GO::performanceScores} and the additional column \code{gene}
(the reference ientifier).
}
\description{
Computes the performance scores of protein function predictions.
}
