% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{testBestBlastF1Score}
\alias{testBestBlastF1Score}
\title{Function testing \code{MapMan2GO::bestBlastF1Score}.}
\usage{
testBestBlastF1Score()
}
\value{
\code{TRUE} if and only if all tests pass.
}
\description{
Function testing \code{MapMan2GO::bestBlastF1Score}.
}
