% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{testF1Score}
\alias{testF1Score}
\title{Tests the function \code{f1Score}.}
\usage{
testF1Score()
}
\value{
\code{TRUE} if and only if all tests pass successfully.
}
\description{
Tests the function \code{f1Score}.
}
