% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{testPerformanceScores}
\alias{testPerformanceScores}
\title{Tests the function \code{performanceScores}.}
\usage{
testPerformanceScores()
}
\value{
\code{TRUE} if and only if all tests pass successfully.
}
\description{
Tests the function \code{performanceScores}.
}
