% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{addAncestors}
\alias{addAncestors}
\title{Adds all GO Terms that are ancestral to the argument \code{go.terms}.}
\usage{
addAncestors(go.terms, ancestors = list(MF = GOMFANCESTOR, BP = GOBPANCESTOR,
  CC = GOCCANCESTOR), exclude.root = TRUE, root.go = "all")
}
\arguments{
\item{go.terms}{a character vector of GO identifier, e.g.
\code{'GO:006969'}.}

\item{ancestors}{a list of keys ontologies and values the respective
ancestral GO terms. Default is \code{list(MF = GOMFANCESTOR, BP =
GOBPANCESTOR, CC = GOCCANCESTOR)}.}

\item{exclude.root}{boolean indicating wether to exclude the ROOT GO Term
\code{'all'}. Default is \code{TRUE}.}

\item{root.go}{String indicating the ROOT GO Term. Default is \code{'all'}.}
}
\value{
A character vector including the argument \code{go.terms} and all
found ancestors.
}
\description{
Adds all GO Terms that are ancestral to the argument \code{go.terms}.
}
