% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{f1Score}
\alias{f1Score}
\title{Compute the F1-Score for the prediction \code{predicted.annos} as the
harmonic mean of precision and recall.}
\usage{
f1Score(true.annos, predicted.annos)
}
\arguments{
\item{true.annos}{A vector of atomic annotations representing the TRUTH.}

\item{predicted.annos}{A vector of atomic annotations representing the
PREDICTIONS.}
}
\value{
Number - The weighted harmonic mean of precision and recall.
}
\description{
Compute the F1-Score for the prediction \code{predicted.annos} as the
harmonic mean of precision and recall.
}
