% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funks.R
\name{mercatorF1Score}
\alias{mercatorF1Score}
\title{For each row in \c{mercator.annos} measures the respective F1-Score of the
MapMan-Bin annotation based on the reference annotations.}
\usage{
mercatorF1Score(mercator.annos,
  reference.gene.annos = getOption("MapMan2GO.reference.gene.annos",
  ukb.ref.goas), rga.gene.col = getOption("MapMan2GO.rga.gene.col", "V4"),
  rga.anno.col = getOption("MapMan2GO.rga.anno.col", "V2"))
}
\arguments{
\item{mercator.annos}{an instance of \c{base::data.frame} produced by
invoking function \c{MapMan2GO::readMercatorResultTable}.}

\item{reference.gene.annos}{An instance of \c{base::data.frame} holding the
reference annotations for the referene genes, i.e. the TRUTH the mercator
predictions are evaluated against. Default is
\c{getOption('MapMan2GO.reference.gene.annos', ukb.ref.goas)}.}

\item{rga.gene.col}{An integer or string indicating in which column of
\c{reference.gene.annos} to lookup the reference gene identifiers. Default
is \c{getOption("MapMan2GO.rga.gene.col", "V4")}.}

\item{rga.anno.col}{An integer or string indicating in which column of
\c{reference.gene.annos} to lookup the reference gene annotations. Default
is \c{getOption("MapMan2GO.rga.anno.col", "V2")}.}
}
\value{
A numeric vector holding the respective F1-Scores.
}
\description{
For each row in \c{mercator.annos} measures the respective F1-Score of the
MapMan-Bin annotation based on the reference annotations.
}
